% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWRandMatrices.R
\name{compute.adjacency.matrix}
\alias{compute.adjacency.matrix}
\title{Computes the adjacency matrix of a multiplex network}
\usage{
compute.adjacency.matrix(x,delta = 0.5)
}
\arguments{
\item{x}{A \code{Multiplex} object describing a multiplex network generated
by the function \code{create.multiplex}.}

\item{delta}{A numeric value between 0 and 1. It sets the probability
of performing inter-layer versus intra-layer transitions. It is set by
default to 0.5. See more details below.}
}
\value{
A square sparse adjacency matrix created with the \code{Matrix}
package.
}
\description{
\code{compute.adjacency.matrix} is a function to compute the adjacency
matrix of a multiplex network provided as a \code{Multiplex} object.
}
\details{
The parameter \code{delta} sets the probability to change between
layers at the next step. If delta = 0, the particle will always remain
in the same layer after a non-restart iteration. On the other hand, if
delta = 1, the particle will always change between layers, therefore
not following the specific edges of each layer.
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(m1,m2)
compute.adjacency.matrix(multiObject)

}
\seealso{
\code{\link{create.multiplex},\link{normalize.multiplex.adjacency},
\link{compute.transition.matrix}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
