% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMultiHetObjects.R
\name{create.multiplex}
\alias{create.multiplex}
\alias{create.multiplex.default}
\title{Create multiplex graphs from individual networks}
\usage{
create.multiplex(...)

\method{create.multiplex}{default}(LayersList, ...)
}
\arguments{
\item{...}{Further arguments passed to \code{create.multiplex}}

\item{LayersList}{A list containing igraph objects describing monoplex 
networks in every element. We recommend to give names to the different 
networks (igraph objects).}
}
\value{
A Multiplex object. It contains a list of the different graphs
integrating the multiplex network, the names and number of its nodes and the
number of layers.
}
\description{
\code{create.multiplex} is a function to create a multiplex network
(\code{Multiplex} object) from a list of individual networks defined as
\code{igraph} objects. See more details about multiplex networks below.
If just one network is provided, a Multiplex object with one layer is
therefore created (A monoplex network).
}
\details{
A multiplex network is a collection of layers (monoplex networks)
sharing the same nodes, but in which the edges represent relationships of
different nature. At least a list with one element, an igraph object, should
be provided.
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(list(m1=m1,m2=m2))

}
\seealso{
\code{\link{create.multiplexHet},\link{isMultiplex}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
