% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InduceNetworks.R
\name{create.multiplexHetNetwork.topResults}
\alias{create.multiplexHetNetwork.topResults}
\title{Creates a Network with the top results of the Random Walk with restart on
a Multiplex and Heterogeneous Network}
\usage{
create.multiplexHetNetwork.topResults(RWRMH_Results_Object,
    MultiplexHetObject, bipartite_relations, bipartite_name, k=25)
}
\arguments{
\item{RWRMH_Results_Object}{A \code{RWRMH_Results} object generated by the
function \code{Random.Walk.Restart.MultiplexHet} representing the results
of the Random Ralk with restart on the multiplex and heterogeneous network
described in the following argument.}

\item{MultiplexHetObject}{A \code{MultiplexHet} object generated by the
function \code{create.multiplexHet} representing a multiplex and
heterogeneous network.}

\item{bipartite_relations}{A data frame containing the relationships (bipartite
interactions) between the nodes of the first multiplex network and the nodes 
of the second multiplex of the heterogeneous system. The data frame should 
contain two or three columns: the first one with the nodes of the multiplex 
network; the second one with the nodes of the second network. 
The third one is not mandatory and it should contain the weights. Every node 
should be present in their corresponding multiplex network.}

\item{bipartite_name}{A vector containing the name for the bipartite
relations to be integrated as part of the resulting network. It is included
as an attribute for all the bipartite edges of the resulting network. It's
optional and its default value is "bipartiteRelations".}

\item{k}{A numeric value between 1 and 200. It is the number of top ranked
nodes to be included in the resulting multiplex network.}
}
\value{
An \code{igraph} object containing the top \code{k} ranked
multiplex nodes and the top \code{k} ranked second network nodes in the
Random Walk with Restart on a Multiplex and Heterogeneous network algorithm.
We include all the possible types of interactions between pairs of
nodes according to the different layers of the multiplex network, the
bipartite interactions and the second network type of interactions.
}
\description{
\code{create.multiplexHetNetwork.topResults} is a function to create a
network from the top results of the Random Walk with Restart on Multiplex and
Heterogeneous networks algorithm (a \code{RWRMH_Results} object).
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject_1 <- create.multiplex(list(m1=m1,m2=m2))
h1 <- igraph::graph(c("A","C","B","E","E","D","E","C"), directed = FALSE)
multiObject_2 <- create.multiplex(list(h1=h1))
bipartite_relations <- data.frame(m=c(1,3),h=c("A","E"))
multiHetObject <- 
    create.multiplexHet(multiObject_1, multiObject_2,bipartite_relations)
MultiHetTranMatrix <- compute.transition.matrix(multiHetObject)
Multiplex1_Seeds <- c(1)
Multiplex2_Seeds <- c("E")
RWR_MultiHetResults <- 
    Random.Walk.Restart.MultiplexHet(MultiHetTranMatrix, multiHetObject,
        Multiplex1_Seeds,Multiplex2_Seeds)
create.multiplexHetNetwork.topResults(RWR_MultiHetResults,multiHetObject,
    bipartite_relations)

}
\seealso{
\code{\link{create.multiplexHet},
\link{isRWRMH_Results}, \link{Random.Walk.Restart.MultiplexHet}
\link{create.multiplexNetwork.topResults}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
