% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InduceNetworks.R
\name{create.multiplexNetwork.topResults}
\alias{create.multiplexNetwork.topResults}
\title{Creates a Network with the top results of the Random Walk with restart on
a Multiplex Network}
\usage{
create.multiplexNetwork.topResults(RWRM_Result_Object,
    MultiplexObject,k=25)
}
\arguments{
\item{RWRM_Result_Object}{A \code{RWRM_Results} object generated by the
function \code{Random.Walk.Restart.Multiplex} representing the results
of the Random Ralk with restart on the multiplex network described in the
following argument.}

\item{MultiplexObject}{A \code{Multiplex} object generated by the
function \code{create.multiplex} representing a multiplex network.}

\item{k}{A numeric value between 1 and 200. It is the number of top ranked
nodes to be included in the resulting multiplex network.}
}
\value{
An \code{igraph} object containing the top \code{k} ranked
multiplex nodes in the Random Walk with Restart on a Multiplex network
algorithm. We include all the possible types of interactions between pairs of
nodes according to the different layers of the multiplex network.
}
\description{
\code{create.multiplexNetwork.topResults} is a function to create a network
from the top results of the Random Walk with Restart on Multiplex networks
algorithm (a \code{RWRM_Results} object).
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(list(m1=m1,m2=m2))
AdjMatrix <- compute.adjacency.matrix(multiObject)
AdjMatrixNorm <- normalize.multiplex.adjacency(AdjMatrix)
Seed <- c(1)
RWR_MultiResults <- 
    Random.Walk.Restart.Multiplex(AdjMatrixNorm, multiObject, Seed)
create.multiplexNetwork.topResults(RWR_MultiResults,multiObject)

}
\seealso{
\code{\link{create.multiplex}, \link{Random.Walk.Restart.Multiplex}
\link{isRWRM_Results}, \link{create.multiplexHetNetwork.topResults}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
