% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Disease_Network}
\alias{Disease_Network}
\title{A disease-disease similarity network.}
\format{An igraph object containing 28246 binary relationships between 6947
diseases.}
\source{
Valdeolivas, A., Tichit, L., Navarro, C., Perrin, S., Odelin, G.,
Levy, N., … Baudot, A. (2017). Random Walk With Restart On Multiplex And
Heterogeneous Biological Networks. bioRxiv, 1–31.
https://doi.org/10.1101/134734
\url{https://www.biorxiv.org/content/early/2017/08/30/134734}
}
\usage{
data(Disease_Network)
}
\description{
An igraph object containing a disease-disease similarity network. 
The network is obtained as described in the article
cited in the source section.
}
\keyword{datasets}
