% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneDiseaseRelations}
\alias{GeneDiseaseRelations}
\title{Diseases and their causative genes}
\format{A data frame with 4496 rows and 2 variables:
\describe{
    \item{hgnc_symbol}{Gene name, in HGNC format}
    \item{mim_morbid}{Disease id, in mim code}
}}
\source{
Valdeolivas, A., Tichit, L., Navarro, C., Perrin, S., Odelin, G.,
Levy, N., … Baudot, A. (2017). Random Walk With Restart On Multiplex And
Heterogeneous Biological Networks. bioRxiv, 1–31.
https://doi.org/10.1101/134734
\url{https://www.biorxiv.org/content/early/2017/08/30/134734}
}
\usage{
data(GeneDiseaseRelations)
}
\description{
A dataset containing some diseases and their causative genes.
The dataset is obtained as described in the article
cited in the source section.
}
\keyword{datasets}
