% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{is.multiplex.het}
\alias{is.multiplex.het}
\title{Is this R object a Multiplex Heterogeneous object?}
\usage{
is.multiplex.het(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A logical constant, \code{TRUE} if argument \code{x} is
a MultiplexHet object.
}
\description{
A Multiplex Heterogeneous object is an R object generated as the result
of calling the function \code{create.multiplexHet}
}
\details{
\code{is.multiplex.het(x)} checks whether an R object is MultiplexHet
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(m1,m2)
h1 <- igraph::graph(c("A","C","B","E","E","D","E","C"), directed = FALSE)
bipartite_relations <- data.frame(m=c(1,3),h=c("A","E"))
multiHetObject <- create.multiplexHet(multiObject,
    h1,bipartite_relations)
is.multiplex.het(multiHetObject)
is.multiplex.het(h1)

}
\seealso{
\code{\link{create.multiplexHet}},
\code{\link{is.multiplex}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
