% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{is.multiplex}
\alias{is.multiplex}
\title{Is this R object a Multiplex object?}
\usage{
is.multiplex(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A logical constant, \code{TRUE} if argument \code{x} is
a Mutiplex object.
}
\description{
A Multiplex object is an R object generated as the result of calling
the function \code{create.multiplex}
}
\details{
\code{is.multiplex(x)} checks whether an R object is Multiplex.
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(m1,m2)
is.multiplex(multiObject)
is.multiplex(m1)

}
\seealso{
\code{\link{create.multiplex}}, \code{\link{is.multiplex.het}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
