% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{is.RWRMH.Results}
\alias{is.RWRMH.Results}
\title{Is this R object a RWR on Multiplex-Heterogeneous object (Results of the
RWR-MH)?}
\usage{
is.RWRMH.Results(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A logical constant, \code{TRUE} if argument \code{x} is
a RWRMH_Results object.
}
\description{
A RWR on Multiplex Heterogeneous object is an R object generated as the
result of calling the function \code{Random.Walk.Restart.MultiplexHet}
(Results of the RWR-MH)
}
\details{
\code{is.RWRMH.Results(x)} checks whether an R object is RWRMH_Results
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(m1,m2)
h1 <- igraph::graph(c("A","C","B","E","E","D","E","C"), directed = FALSE)
bipartite_relations <- data.frame(m=c(1,3),h=c("A","E"))
multiHetObject <- create.multiplexHet(multiObject,
    h1,bipartite_relations)
MultiHetTranMatrix <- compute.transition.matrix(multiHetObject)
Multiplex_Seeds <- c(1)
SecondNet_Seeds <- c("E")
RWR_MultiHetResults <- Random.Walk.Restart.MultiplexHet(MultiHetTranMatrix,
    multiHetObject,Multiplex_Seeds,SecondNet_Seeds)
is.RWRMH.Results(RWR_MultiHetResults)
is.RWRMH.Results(m1)

}
\seealso{
\code{\link{Random.Walk.Restart.MultiplexHet}},
\code{\link{is.RWRM.Results}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
