% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckObjects.R
\name{isMultiplex}
\alias{isMultiplex}
\title{Is this R object a Multiplex object?}
\usage{
isMultiplex(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A logical constant, \code{TRUE} if argument \code{x} is
a Mutiplex object.
}
\description{
A Multiplex object is an R object generated as the result of calling
the function \code{create.multiplex}
+
\code{isMultiplex(x)} checks whether an R object is Multiplex.
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(m1,m2)
isMultiplex(multiObject)
isMultiplex(m1)

}
\seealso{
\code{\link{create.multiplex}}, \code{\link{isMultiplexHet}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
