% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckObjects.R
\name{isRWRM_Results}
\alias{isRWRM_Results}
\title{Is this R object a RWR on Multiplex object (Results of the RWR-M)?}
\usage{
isRWRM_Results(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A logical constant, \code{TRUE} if argument \code{x} is
a RWRM_Results object.
}
\description{
A RWR on Multiplex object is an R object generated as the result
of calling the function \code{Random.Walk.Restart.Multiplex}
(Results of the RWR-M)
}
\details{
\code{isRWRM_Results(x)} checks whether an R object is RWRM_Results
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(list(m1=m1,m2=m2))
AdjMatrix <- compute.adjacency.matrix(multiObject)
AdjMatrixNorm <- normalize.multiplex.adjacency(AdjMatrix)
Seed <- c(1)
RWR_MultiResults <- 
    Random.Walk.Restart.Multiplex(AdjMatrixNorm, multiObject,Seed)
isRWRM_Results(RWR_MultiResults)
isRWRM_Results(m1)
}
\seealso{
\code{\link{Random.Walk.Restart.Multiplex}},
\code{\link{isRWRMH_Results}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
