% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckObjects.R
\name{isRWRMH_Results}
\alias{isRWRMH_Results}
\title{Is this R object a RWR on Multiplex-Heterogeneous object (Results of the
RWR-MH)?}
\usage{
isRWRMH_Results(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A logical constant, \code{TRUE} if argument \code{x} is
a RWRMH_Results object.
}
\description{
A RWR on Multiplex Heterogeneous object is an R object generated as the
result of calling the function \code{Random.Walk.Restart.MultiplexHet}
(Results of the RWR-MH)
}
\details{
\code{isRWRMH_Results(x)} checks whether an R object is RWRMH_Results
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject_1 <- create.multiplex(list(m1=m1,m2=m2))
h1 <- igraph::graph(c("A","C","B","E","E","D","E","C"), directed = FALSE)
multiObject_2 <- create.multiplex(list(h1=h1))
bipartite_relations <- data.frame(m=c(1,3),h=c("A","E"))
multiHetObject <-
    create.multiplexHet(multiObject_1,multiObject_2,bipartite_relations)
MultiHetTranMatrix <- compute.transition.matrix(multiHetObject)
Multiplex1_Seeds <- c(1)
Multiplex2_Seeds <- c("E")
RWR_MultiHetResults <- 
    Random.Walk.Restart.MultiplexHet(MultiHetTranMatrix,multiHetObject,
    Multiplex1_Seeds ,Multiplex2_Seeds)
isRWRMH_Results(RWR_MultiHetResults)
isRWRMH_Results(m1)

}
\seealso{
\code{\link{Random.Walk.Restart.MultiplexHet}},
\code{\link{isRWRM_Results}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
