% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWRandMatrices.R
\name{normalize.multiplex.adjacency}
\alias{normalize.multiplex.adjacency}
\title{Computes column normalization of an adjacency matrix}
\usage{
normalize.multiplex.adjacency(x)
}
\arguments{
\item{x}{A \code{Matrix} object describing an adjacency matrix of a network.}
}
\value{
A square sparse column normalized matrix created with the
\code{Matrix} package.
}
\description{
\code{normalize.multiplex.adjacency} is a function to compute the column
normalization of a sparse matrix of the package \code{Matrix}.
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(list(m1=m1,m2=m2))
AdjMatrix <- compute.adjacency.matrix(multiObject)
normalize.multiplex.adjacency(AdjMatrix)

}
\seealso{
\code{\link{compute.adjacency.matrix},
\link{Random.Walk.Restart.Multiplex}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
