% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pathway_Network}
\alias{Pathway_Network}
\title{A pathway network (Pathway network)}
\format{
An igraph object containing 62602 binary interactions between 3533
proteins
}
\source{
Valdeolivas, A., Tichit, L., Navarro, C., Perrin, S., Odelin, G.,
Levy, N., … Baudot, A. (2017). Random Walk With Restart On Multiplex And
Heterogeneous Biological Networks. bioRxiv, 1–31.
https://doi.org/10.1101/134734
\url{https://www.biorxiv.org/content/early/2017/08/30/134734}
}
\usage{
data(Pathway_Network)
}
\description{
An igraph object containing a Pathway network. 
The network is obtained as described in the article 
cited in the source section. However, it was reduced in such a way
that only contains genes/proteins expressed in the adipose tissue.
}
\keyword{datasets}
