% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWRandMatrices.R
\name{row.normalize.matrix}
\alias{row.normalize.matrix}
\title{Computes row normalization of a matrix}
\usage{
row.normalize.matrix(x)
}
\arguments{
\item{x}{A \code{Matrix} object describing a matrix.}
}
\value{
A sparse column normalized matrix created with the
\code{Matrix} package.
}
\description{
\code{row.normalize.matrix} is a function to compute the row
normalization of a sparse matrix of the package \code{Matrix}.
}
\examples{
m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(list(m1=m1,m2=m2))
AdjMatrix <- compute.adjacency.matrix(multiObject)
row.normalize.matrix(AdjMatrix)

}
\seealso{
\code{\link{compute.adjacency.matrix},
\link{Random.Walk.Restart.Multiplex}}
}
\author{
Alberto Valdeolivas Urbelz \email{alvaldeolivas@gmail.com}
}
